+++ Frends.Redis.KeyExists/Frends.Redis.KeyExists/Frends.Redis.KeyExists.cs
@@ -5,15 +5,12 @@ using System.ComponentModel;
 using System.Threading.Tasks;
 using Definitions;
 using Helpers;
-using StackExchange.Redis;
 
 /// <summary>
 /// Main class of the Task.
 /// </summary>
 public static class Redis
 {
-    private static IConnectionMultiplexer redis;
-
     /// <summary>
     /// This is Task to check if Redis key exists.
     /// [Documentation](https://tasks.frends.com/tasks/frends-tasks/Frends.Redis.KeyExists).
@@ -22,11 +19,14 @@ public static class Redis
     /// <param name="options">Exception settings.</param>
     /// <param name="connection">Connection info.</param>
     /// <returns>Object { bool Success, Error error }.</returns>
-    public static async Task<Result> KeyExists([PropertyTab] Input input, [PropertyTab] Options options, [PropertyTab] Connection connection)
+    public static async Task<Result> KeyExists(
+        [PropertyTab] Input input,
+        [PropertyTab] Options options,
+        [PropertyTab] Connection connection)
     {
         try
         {
-            redis = await ConnectionMultiplexer.ConnectAsync(connection.ConnectionString);
+            await using var redis = await ConnectionHandler.GetConnectionAsync(connection).ConfigureAwait(false);
             var db = redis.GetDatabase();
             var exists = await db.KeyExistsAsync(input.Key);
             return new Result(exists);
@@ -34,10 +34,6 @@ public static class Redis
         catch (Exception ex)
         {
             return ErrorHandler.Handle(ex, options.ThrowErrorOnFailure, options.ErrorMessageOnFailure);
-        }
-        finally
-        {
-            redis?.Dispose();
         }
     }
 }