+++ Frends.Redis.SetValue/Frends.Redis.SetValue/Definitions/ConnectionMethod.cs
@@ -1,4 +1,4 @@
-﻿namespace Frends.Redis.GetValue.Definitions;
+﻿namespace Frends.Redis.SetValue.Definitions;
 
 /// <summary>
 /// Method to use to connect to Redis