+++ Frends.Redis.SetValue/CHANGELOG.md
@@ -1,16 +1,22 @@
 # Changelog
 
 ## [2.0.0] - 2026-03-03
+
 ### Added
+
 - Add support for more Redis value types (string, hash, list, set)
 - **[Breaking Change]** Rename Input.Value to Input.StringValue to distinguish between values of different types.
 - Add option to define if collection-like values will be appended or overwritten.
-- Add support for new connection methods
-- 
+- Add support for Microsoft Entra ID
+
 ## [1.1.0] - 2025-07-31
+
 ### Changed
+
 - Make the default error message empty by default
 
 ## [1.0.0] - 2025-05-13
+
 ### Changed
+
 - Initial implementation